requirements.txt
python-telegram-bot==20.7
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0 